refactor: shared components directory
