+++ packages/web/src/shared/components/DropdownMenu.tsx
@@ -1,6 +1,6 @@
 import { FocusEvent, ReactNode, useRef, useState } from 'react';
 import { Link, LinkProps } from 'react-router-dom';
-import useCssId from './cssId';
+import useCssId from '../cssId';
 import { Icon } from './Icon';
 
 export interface DropdownProps {
